rename api package to client for clarity
