+++ client/requester.go
@@ -1,4 +1,4 @@
-package api
+package client
 
 import (
 	"fmt"
